workspace test refactoring
